requirements.txt
fastapi
uvicorn[standard]
redis
numpy
librosa
psycopg2-binary
pydantic-settings